# Optimize metrics wrapper