Fairly certain we can optimise it further, but at least now it's not cripplingly slow. Open to suggestions on what else can be improved. The metrics wrapper is now ~x50 faster but still twice as slow as the `ApolloTracing` wrapper.

Before: 

```
[info] Benchmark                                      Mode  Cnt    Score    Error  Units
[info] NestedZQueryBenchmark.apolloTracingBenchmark  thrpt    5  202.018 ± 21.100  ops/s
[info] NestedZQueryBenchmark.metricsBenchmark        thrpt    5    5.500 ±  0.258  ops/s
[info] NestedZQueryBenchmark.noWrappersBenchmark     thrpt    5  511.841 ±  8.066  ops/s
```

After
```
[info] Benchmark                                      Mode  Cnt    Score   Error  Units
[info] NestedZQueryBenchmark.apolloTracingBenchmark  thrpt    5  199.308 ± 3.427  ops/s
[info] NestedZQueryBenchmark.metricsBenchmark        thrpt    5  117.211 ± 3.722  ops/s
[info] NestedZQueryBenchmark.noWrappersBenchmark     thrpt    5  515.891 ± 7.015  ops/s
```

If we fork the writing of metrics, then the `metricsBenchmark` becomes faster than the `ApolloTracing` wrapper:

```
[info] Benchmark                                Mode  Cnt    Score   Error  Units
[info] NestedZQueryBenchmark.metricsBenchmark  thrpt    5  213.406 ± 4.114  ops/s
```